remove core submodule
