--- artist/core/__init__.py
@@ -1,11 +1,0 @@
-"""Bundle all classes responsible for the core functions in ``ARTIST``."""
-
-from artist.optimization.kinematics_reconstructor import KinematicsReconstructor
-from artist.optimization.motor_position_optimizer import MotorPositionsOptimizer
-from artist.optimization.surface_reconstructor import SurfaceReconstructor
-
-__all__ = [
-    "KinematicsReconstructor",
-    "SurfaceReconstructor",
-    "MotorPositionsOptimizer",
-]
